initial work in setting an internal DB table for triggers managment
