+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 CHANGELOG
 ===
+
 **NEXT: (XXXX/XX/XX)**
 
 + BETHA, app can now store INT tags values from MODBUS TCP/IP slave, to LOG,LASTVALUE tables of SQLITE3,MYSQL databases.
-
-    - LASTVALUE tables added, where only last value of a tag is saved
+    - NEW FEATURES:LASTVALUE tables added, where only last value of a tag is saved
+    - BUGFIX: not trying to store data when communications were completely wrong
+    - CLEANING AND POLISHING: old code not needed removed
 
 **v0.0.2: (2016/12/13)**
