+++ README.md
@@ -19,7 +19,7 @@ prointegra-OPC is a C++ application to take/send tags from/to industrial automat
 + QT 5 and it's drivers for databases. (MySQL, SQLite3)
     - under ubuntu xenial (16.04): ``` sudo apt-get install libqt5sql-mysql libqt5sql-sqlite ```
 + use of [gStools](https://github.com/gentooza/gStools) for tags conversions (already included in /lib folder)
-+ use of pugiXML for configurations files (already cinluded in /lib folder, v1.7)
++ use of [pugiXML](https://github.com/zeux/pugixml) for configurations files (already inluded in /lib folder, v1.7)
 
 ##Docs
