+++ ROADMAP.md
@@ -3,6 +3,7 @@ ROADMAP
 + simple terminal gui to manage program
 + differents saving methods to database: by time, by tag trigger
 + write tags from database values
++ database triggers table and conversation with clients to read & write data by triggers
 + MORE databases support
     - TDS support
 + MORE tags type support
