+++ config.cpp
@@ -182,6 +182,7 @@ int ConfigParser::retrieveTablesParams(pugi::xml_node* db, int dbNumber, int num
       retrieveCharAttr(&table,&tablesParams[dbNumber][i].tbName,"name");
       retrieveCharAttr(&table,&tablesParams[dbNumber][i].tbTrigger,"tagReadTrigger");
       retrieveIntAttr(&table,&tablesParams[dbNumber][i].tbTriggerTime,"timeReadTrigger");
+      retrieveCharAttr(&table,&tablesParams[dbNumber][i].tbWTrigger,"tagWriteTrigger");
       retrieveCharAttr(&table,&tablesParams[dbNumber][i].tbType,"type");
       //tags
       tablesParams[dbNumber][i].numFields = retrieveNumberofNodes(&table,"tag");
