+++ prointegra-OPC.pro
@@ -15,6 +15,9 @@ HEADERS += prointegraOPC.h
 HEADERS += qtdatabase.h
 HEADERS += commDaemon.h
 HEADERS += piDatabase.h
+HEADERS += piTable.h
+#HEADERS += piTriggersTable.h
+HEADERS += piDataTable.h
 HEADERS += piComm.h
 HEADERS += iniConfigurator.h
 HEADERS += config.h
@@ -26,6 +29,9 @@ SOURCES += prointegraOPC.cpp
 SOURCES += qtdatabase.cpp
 SOURCES += commDaemon.cpp
 SOURCES += piDatabase.cpp
+SOURCES += piTable.cpp
+#SOURCES += piTriggersTable.cpp
+SOURCES += piDataTable.cpp
 SOURCES += piComm.cpp
 SOURCES += iniConfigurator.cpp
 SOURCES += config.cpp
